fix up middleware for refID validation
