+++ src/main/java/org/breedinginsight/brapps/importer/services/processors/experiment/appendoverwrite/ExpUnitContextService.java
@@ -1,7 +1,6 @@
 package org.breedinginsight.brapps.importer.services.processors.experiment.appendoverwrite;
 
 import io.micronaut.context.annotation.Property;
-import io.micronaut.http.server.exceptions.InternalServerException;
 import lombok.extern.slf4j.Slf4j;
 import org.brapi.client.v2.model.exceptions.ApiException;
 import org.brapi.v2.model.BrAPIExternalReference;
@@ -12,9 +11,8 @@ import org.breedinginsight.brapps.importer.model.response.PendingImportObject;
 import org.breedinginsight.brapps.importer.services.ExternalReferenceSource;
 import org.breedinginsight.brapps.importer.services.processors.experiment.ExperimentUtilities;
 import org.breedinginsight.brapps.importer.services.processors.experiment.appendoverwrite.model.ExpUnitContext;
-import org.breedinginsight.brapps.importer.services.processors.experiment.model.ExpImportProcessErrorConstants;
+import org.breedinginsight.brapps.importer.services.processors.experiment.model.ExpImportProcessConstants;
 import org.breedinginsight.brapps.importer.services.processors.experiment.model.ImportContext;
-import org.breedinginsight.model.Program;
 import org.breedinginsight.utilities.Utilities;
 
 import javax.inject.Inject;
@@ -79,7 +77,7 @@ public class ExpUnitContextService {
                 missingIds.removeAll(fetchedIds);
 
                 // throw error reporting any reference IDs with no corresponding stored unit in the brapi data store
-                throw new IllegalStateException("Observation Units not found for ObsUnitId(s): " + String.join(ExpImportProcessErrorConstants.COMMA_DELIMITER, missingIds));
+                throw new IllegalStateException("Observation Units not found for ObsUnitId(s): " + String.join(ExpImportProcessConstants.COMMA_DELIMITER, missingIds));
             }
 
             return pendingUnitById;
