+++ src/main/java/org/breedinginsight/brapps/importer/services/processors/experiment/appendoverwrite/middleware/ValidateAllRowsHaveIDs.java
@@ -1,6 +1,8 @@
 package org.breedinginsight.brapps.importer.services.processors.experiment.appendoverwrite.middleware;
 
+import lombok.extern.slf4j.Slf4j;
 import org.breedinginsight.brapps.importer.model.imports.experimentObservation.ExperimentObservation;
+import org.breedinginsight.brapps.importer.services.processors.experiment.ExperimentUtilities;
 import org.breedinginsight.brapps.importer.services.processors.experiment.appendoverwrite.middleware.ExpUnitMiddleware;
 import org.breedinginsight.brapps.importer.services.processors.experiment.model.ExpUnitMiddlewareContext;
 import org.breedinginsight.brapps.importer.services.processors.experiment.model.MiddlewareError;
@@ -8,13 +10,12 @@ import org.breedinginsight.brapps.importer.services.processors.experiment.model.
 import java.util.HashSet;
 import java.util.Set;
 
+@Slf4j
 public class ValidateAllRowsHaveIDs extends ExpUnitMiddleware {
-    private boolean hasAllReferenceUnitIds = true;
-    private boolean hasNoReferenceUnitIds = true;
-
     @Override
     public boolean process(ExpUnitMiddlewareContext context) {
-        context.setReferenceOUIds(collateReferenceOUIds(context));
+
+        context.getExpUnitContext().setReferenceOUIds(ExperimentUtilities.collateReferenceOUIds(context));
         return processNext(context);
     }
 
@@ -27,27 +28,5 @@ public class ValidateAllRowsHaveIDs extends ExpUnitMiddleware {
         return compensatePrior(context, error);
     }
 
-    private Set<String> collateReferenceOUIds(ExpUnitMiddlewareContext context) {
-        Set<String> referenceOUIds = new HashSet<>();
-        for (int rowNum = 0; rowNum < context.getImportContext().getImportRows().size(); rowNum++) {
-            ExperimentObservation importRow = (ExperimentObservation) context.getImportContext().getImportRows().get(rowNum);
 
-            // Check if ObsUnitID is blank
-            if (importRow.getObsUnitID() == null || importRow.getObsUnitID().isBlank()) {
-                hasAllReferenceUnitIds = false;
-            } else if (referenceOUIds.contains(importRow.getObsUnitID())) {
-
-                // Throw exception if ObsUnitID is repeated
-                this.compensate(context, new MiddlewareError(()->{
-                    throw new IllegalStateException("ObsUnitId is repeated: " + importRow.getObsUnitID());
-                }));
-
-            } else {
-                // Add ObsUnitID to referenceOUIds
-                referenceOUIds.add(importRow.getObsUnitID());
-                hasNoReferenceUnitIds = false;
-            }
-        }
-        return referenceOUIds;
-    }
 }
