+++ src/main/java/org/breedinginsight/brapps/importer/services/processors/experiment/appendoverwrite/middleware/commit/CreateBrAPITrials.java
@@ -1,0 +1,4 @@
+package org.breedinginsight.brapps.importer.services.processors.experiment.appendoverwrite.middleware.commit;
+
+public class CreateBrAPITrials {
+}
