--- src/main/java/org/breedinginsight/brapps/importer/services/processors/experiment/appendoverwrite/middleware/postBrAPIData/CreateBrAPITrials.java
@@ -1,4 +1,0 @@
-package org.breedinginsight.brapps.importer.services.processors.experiment.appendoverwrite.middleware.postBrAPIData;
-
-public class CreateBrAPITrials {
-}
