+++ src/main/java/org/breedinginsight/brapps/importer/services/processors/experiment/model/ExpUnitMiddlewareContext.java
@@ -4,10 +4,9 @@ import lombok.*;
 import org.breedinginsight.brapps.importer.services.processors.experiment.appendoverwrite.model.ExpUnitContext;
 import org.breedinginsight.brapps.importer.services.processors.experiment.create.model.PendingData;
 
-@Getter
-@Setter
+@Data
+@Builder
 public class ExpUnitMiddlewareContext {
-
     private ImportContext importContext;
     private ExpUnitContext expUnitContext;
     private PendingData pendingData;
